move json pack into core
